Move some modules and document StateRegistry
