+++ src/components/mod.rs
@@ -9,9 +9,9 @@ use crate::{
 };
 
 pub mod archive;
-pub mod bound;
 pub mod boundary;
 pub mod control_flow;
+pub mod cro;
 pub mod diversity;
 pub mod evaluation;
 pub mod generative;
