+++ src/conditions/cro.rs
@@ -2,7 +2,7 @@ use eyre::{eyre, WrapErr};
 use serde::{Deserialize, Serialize};
 
 use crate::{
-    component::ExecResult, components::bound::cro::ChemicalReaction, conditions::Condition,
+    component::ExecResult, components::cro::ChemicalReaction, conditions::Condition,
     population::IntoSingleRef, Problem, State,
 };
