+++ src/prelude.rs
@@ -1,7 +1,9 @@
 //! The MAHF prelude imports the most relevant modules, structs and traits you may need for experiments.
 
 pub use crate::{
-    components::{self, *},
+    components::{
+        self, evaluation, initialization, mapping, mutation, recombination, replacement, selection,
+    },
     conditions,
     heuristics::*,
     logging,
