+++ src/state/registry/error.rs
